extract time helpers (toTime, clockTime, getTimezoneTime, getTimezoneOffset) to lib/time.ts
